+++ NEWS.md
@@ -1,5 +1,8 @@
 # 0.21.3.9000
 
+* If a study has no features table for a modelID, a study can still have
+linkouts to external resources in the results table, as long as the linkouts
+only refer to the featureID column.
 
 # 0.21.3